requirements.txt
requests==2.28.1
tenacity==8.2.3
python-dotenv==1.0.0
pytest==7.4.0